+++ projects/igniteui-angular-i18n/src/i18n/IT/action-strip-resources.ts
@@ -1,9 +1,7 @@
-import { IActionStripResourceStrings } from 'igniteui-angular';
+import { IActionStripResourceStrings, convertToIgxResource } from 'igniteui-angular';
+import { ActionStripResourceStringsIT as AActionStripResourceStrings } from 'igniteui-i18n-resources';
 
-// only use `satisfies` operator so export is typed by its schema
 /**
  * Italian resource strings for IgxActionStrip
  */
-export const ActionStripResourceStringsIT = {
-    igx_action_strip_button_more_title: 'Altro'
-} satisfies MakeRequired<IActionStripResourceStrings>;
+export const ActionStripResourceStringsIT: IActionStripResourceStrings = convertToIgxResource(AActionStripResourceStrings);