+++ projects/igniteui-angular-i18n/src/i18n/IT/date-picker-resources.ts
@@ -1,10 +1,7 @@
-import { IDatePickerResourceStrings } from 'igniteui-angular';
+import { IDatePickerResourceStrings, convertToIgxResource } from 'igniteui-angular';
+import { DatePickerResourceStringsIT as ADatePickerResourceStrings } from 'igniteui-i18n-resources';
 
-// only use `satisfies` operator so export is typed by its schema
 /**
  * Italian resource strings for IgxDatePicker
  */
-export const DatePickerResourceStringsIT = {
-    igx_date_picker_change_date: 'Modifica data',
-    igx_date_picker_choose_date: 'Scegli data'
-} satisfies MakeRequired<IDatePickerResourceStrings>;
+export const DatePickerResourceStringsIT: IDatePickerResourceStrings = convertToIgxResource(ADatePickerResourceStrings);