+++ projects/igniteui-angular-i18n/src/i18n/JA/tree-resources.ts
@@ -1,10 +1,7 @@
-import { ITreeResourceStrings } from 'igniteui-angular';
+import { ITreeResourceStrings, convertToIgxResource } from 'igniteui-angular';
+import { TreeResourceStringsJA as ATreeResourceStrings } from 'igniteui-i18n-resources';
 
-// only use `satisfies` operator so export is typed by its schema
 /**
  * Japanese resource strings for IgxTree
  */
-export const TreeResourceStringsJA = {
-    igx_expand: '展開',
-    igx_collapse: '縮小'
-} satisfies MakeRequired<ITreeResourceStrings>;
+export const TreeResourceStringsJA: ITreeResourceStrings = convertToIgxResource(ATreeResourceStrings);