+++ projects/igniteui-angular-i18n/src/i18n/RO/chip-resources.ts
@@ -1,10 +1,7 @@
-import { IChipResourceStrings } from 'igniteui-angular';
+import { IChipResourceStrings, convertToIgxResource } from 'igniteui-angular';
+import { ChipResourceStringsRO as AChipResourceStrings } from 'igniteui-i18n-resources';
 
-// only use `satisfies` operator so export is typed by its schema
 /**
  * Romanian resource strings for IgxChip
  */
-export const ChipResourceStringsRO = {
-    igx_chip_remove: 'Eliminați cip',
-    igx_chip_select: 'Selectați cip'
-} satisfies MakeRequired<IChipResourceStrings>;
+export const ChipResourceStringsRO: IChipResourceStrings = convertToIgxResource(AChipResourceStrings);