+++ projects/igniteui-angular-i18n/src/i18n/RO/list-resources.ts
@@ -1,10 +1,7 @@
-import { IListResourceStrings } from 'igniteui-angular';
+import { IListResourceStrings, convertToIgxResource } from 'igniteui-angular';
+import { ListResourceStringsRO as AListResourceStrings } from 'igniteui-i18n-resources';
 
-// only use `satisfies` operator so export is typed by its schema
 /**
  * Romanian resource strings for IgxList
  */
-export const ListResourceStringsRO = {
-    igx_list_no_items: 'Nu există articole în listă.',
-    igx_list_loading: 'Se încarcă datele de pe server...'
-} satisfies MakeRequired<IListResourceStrings>;
+export const ListResourceStringsRO: IListResourceStrings = convertToIgxResource(AListResourceStrings);