+++ projects/igniteui-angular-i18n/src/i18n/SV/list-resources.ts
@@ -1,10 +1,7 @@
-import { IListResourceStrings } from 'igniteui-angular';
+import { IListResourceStrings, convertToIgxResource } from 'igniteui-angular';
+import { ListResourceStringsSV as AListResourceStrings } from 'igniteui-i18n-resources';
 
-// only use `satisfies` operator so export is typed by its schema
 /**
  * Swedish resource strings for IgxList
  */
-export const ListResourceStringsSV = {
-    igx_list_no_items: 'Det finns inga objekt i listan.',
-    igx_list_loading: 'Laddar data från servern...'
-} satisfies MakeRequired<IListResourceStrings>;
+export const ListResourceStringsSV: IListResourceStrings = convertToIgxResource(AListResourceStrings);