+++ projects/igniteui-angular/src/lib/core/i18n/tree-resources.ts
@@ -1,9 +1,9 @@
+import { TreeResourceStringsEN as ATreeResourceStrings } from 'igniteui-i18n-core';
+import { convertToIgxResource } from './resources';
+
 export interface ITreeResourceStrings {
     igx_expand?: string;
     igx_collapse?: string;
 }
 
-export const TreeResourceStringsEN: ITreeResourceStrings = {
-    igx_expand: 'Expand',
-    igx_collapse: 'Collapse',
-};
+export const TreeResourceStringsEN: ITreeResourceStrings = convertToIgxResource(ATreeResourceStrings);