+++ projects/igniteui-angular/src/lib/grids/pivot-grid/pivot-grid.component.ts
@@ -70,7 +70,7 @@ import { IgxGridExcelStyleFilteringComponent, IgxExcelStyleColumnOperationsTempl
 import { IgxPivotGridNavigationService } from './pivot-grid-navigation.service';
 import { IgxPivotColumnResizingService } from '../resizing/pivot-grid/pivot-resizing.service';
 import { IgxFlatTransactionFactory, IgxOverlayService, State, Transaction, TransactionService } from '../../services/public_api';
-import { cloneArray, PlatformUtil, resizeObservable } from '../../core/utils';
+import { cloneArray, onResourceChangeHandle, PlatformUtil, resizeObservable } from '../../core/utils';
 import { IgxPivotFilteringService } from './pivot-filtering.service';
 import { DataUtil, GridColumnDataType } from '../../data-operations/data-util';
 import { IFilteringExpressionsTree } from '../../data-operations/filtering-expressions-tree';
@@ -1060,6 +1060,10 @@ export class IgxPivotGridComponent extends IgxGridBaseDirective implements OnIni
             localeId,
             platform,
             _diTransactions);
+        onResourceChangeHandle(this.destroy$, () => {
+            // Since the columns are kinda static, due to assigning DisplayName on init, they need to be regenerated.
+            this.setupColumns();
+        }, this);
     }
 
     /**